// web/api/server/models/FunctionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FunctionDto = {
    /**
     * The name of the function
     */
    name: string;
    /**
     * The runtime environment of the function
     */
    env: string;
    /**
     * The package of the function
     */
    pkg: string;
    /**
     * The entrypoint of the function
     */
    entry?: string;
};